ath79: ar934x: Update dts for current ag71xx driver
for ar9341 mdio1 is always used otherwise we won't be able to control builtin switch.
To get mdio1 work, eth1 should be enabled as it's also a "simple-mfd".
Without enabling eth1, mdio1 won't be probed.
PS: If the device has only one port and you only want to use gmac0,
change compatible of gmac1 to "syscon", "simple-mfd" in dts.

Signed-off-by: Chuanhong Guo <gch981213@gmail.com>